+++ src/dashboard/apps/consent/settings.py
@@ -12,3 +12,19 @@ from django.conf import settings
 # CONSENT_NUMBER_DAYS_END_DATE = None will return 2024-12-31 23:59:59 (if calculated
 # during the year 2024).
 CONSENT_NUMBER_DAYS_END_DATE = getattr(settings, "CONSENT_NUMBER_DAYS_END_DATE", None)
+
+# Control authority contact for consent validation.
+CONSENT_CONTROL_AUTHORITY = getattr(
+    settings,
+    "CONSENT_CONTROL_AUTHORITY",
+    {
+        "name": "<administration name>",
+        "address_1": "<Adresse 1>",
+        "address_2": "<Adresse complement>",
+        "zip_code": "92000",
+        "city": "<city>",
+        "represented_by": "<represented by>",
+        "email": "<email@gouv.fr>",
+    },
+)
+CONSENT_DONE_AT = getattr(settings, "CONSENT_DONE_AT", "<city>")